refactor: fix circular import by extracting SourceDocumentMetadata to separate module
- Move SourceDocumentMetadata from document_store.py to source_metadata.py
- Update imports in both document_store.py and sqlalchemy_document_store.py
- Import SQLAlchemyDocumentStore at top of document_store.py to fix E402
- Update storage __init__.py to export SourceDocumentMetadata
- All 43 document store tests still passing
